finished part 4 2h

+++ Part 4/bloglist/controllers/blogs.js
@@ -1,22 +1,26 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
-
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = request.user
+  console.log(user)
   if(!body.title){
     return response.status(400).end()
   }
   if(!body.url){
     return response.status(400).end()
   }
-
-  const user = await User.findById(body.userId)
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -33,9 +37,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
-
+  console.log(request.token)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === decodedToken.id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }else {
+    return response.status(401).json({ error: 'username invalid' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
